feat: 프로젝트가 없는 경우 대체 이미지를 설정 (CM-144)
