+++ app/globals.css
@@ -45,6 +45,8 @@
   --color-gray-01: oklch(0.69 0 0); /*#9B9B9B*/
   --color-cm: oklch(0.48 0.0525 40.69); /*#795548*/
   --color-cm-100: oklch(0.85 0.0146 39.41); /*#d7cbc7*/
+  --color-cm-300: oklch(0.6471 0.0334 40.8); /*#a1887f*/
+  --color-cm-700: oklch(0.4015 0.0436 37.96); /*#5d4037*/
   --color-cm-900: oklch(0.3 0.0358 30.2); /*#3e2723*/
   --color-cm-light: oklch(0.94 0.0061 43.33); /*#efeae8*/
   --color-cm-green-light: oklch(0.98 0.0067 145.52); /*#f6faf6*/
